backend/requirements.txt
# NoorGuard Ultimate - FastAPI Backend Dependencies
# Python 3.11+ Required

# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-socketio==5.11.1
aiohttp==3.9.1

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Cache
redis==5.0.1
aioredis==2.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# AI & ML
tensorflow-cpu==2.15.0
numpy==1.26.3
scikit-learn==1.4.0
pillow==10.2.0

# Islamic Features
hijri-converter==2.3.1
praytimes==2.1

# Utilities
python-dateutil==2.8.2
python-dotenv==1.0.0
httpx==0.26.0
bcrypt==4.1.2

# Logging & Monitoring
python-json-logger==2.0.7
loguru==0.7.2
sentry-sdk==1.39.1

# File Storage
boto3==1.34.14
aiofiles==23.2.1

# Data Validation
email-validator==2.1.0.post1

# CORS
starlette-cors==0.2.1
